Handle null timestamp or runtime TvMaze fields
In rare cases, the TvMaze database provides null timestamp
or runtime data fields. To make the metadata retrieval script
more fault tolerant, null checks have been added for these values.

When the timestamp field is null, don't use it to generate a
datetime to use in a search for a matching episode.

When the runtime duration field is null, the delta size is
now set to zero minutes. The code which looks for an episode
match has been updated to separate "on time" recordings from
"a little late" recordings to accommodate the possible delta
size of zero.

Refs: #654