Default to last channel used in Manual Record rule
When a user is entering a number of Manual Record rules
in a row, it would be nice if the previously used channel
was remembered and used as the default for the current
rule. For example, if you wanted to enter rules to record
Chicago Med, Chicago Fire, and Chicago P.D., which are
all on the same channel, it would save time on the second
and third rules, to default to the common channel.

To implement this, when a Manual Record rule is entered,
store the channel number such that it can be set as the
default channel for the next Manual Record rule.

Fixes #655