GuessLookupType improvements
To improve the accuracy of GuessLookupType(), the
recording type and recording length are now factored
in. If the recording type is All, Daily, or Weekly,
then it's probably a television show. When there is
no episode number or subtitle, the function used to
assume it's a movie. But for a Manual Record of a
TV show or a movie, there generally isn't an
episode number or subtitle, so this presumption
was frequently wrong. As a last resort, we'll use
the length of the recording as a hint. If a
recording is less than 90 minutes, we'll guess
it's a television show. Otherwise, we'll guess
it's a movie.

Refs #656