Stop logger before exiting mytharchivehelper
The check of the commandline parameters can result in an error message plus an
immediate exit when there is a parameter missing or incorrect.
This message is given with the logging system, but because the logging system
is running on a separate thread the program is being exited while the logging system
is still active. This usually results in a segmentation fault in the logger.
This is now fixed by stopping the logger before exiting.