Add dependencies for mythffmpeg etc.
Add dependency in makefile for mythffmpeg so that it gets
updated whenever there is a new ffmpeg_g.
Same for the other mythff* applications.
This bug surfaced when investigating the failure of the
mytharchivehelper, fixed now in commit 987da09e83ca80181afdd24a1b457faadd1f2c9f.
N.B. The fix is in the mythtv-specific part of an ffmpeg makefile
so possibly the mechanism that is used to update the ffmpeg makefile
also needs to be updated or regenerated.