GuessLookupType improvements
To improve the accuracy of GuessLookupType(), the
recording type is now factored in. If the recording
type is Daily or Weekly, then it's probably a
television show.

The Load() method returns a boolean to indicate
success or failure in reading the database. This
used to be ignored, but now it is checked for
success before using values from the database
read.

Refs #656