Add bounds checking to the DVB NIT/SDT table parsing.
When parsing any of the variable length portions of any of these
tables, ensure that the parsing never reads past the end of the
packet.  Knowing the number of entries in the table up front also
allows for preallocating all the space needed up front, instead of
std::vector making multiple allocation requests as the vector size
grows.