DiSEqC support for SatIP
The "src" value in the SatIP tune command selects the DiSEqC position in the minisatip server.
The default value of 1 is used to select the LNB at the first position and this
is also the correct value when the LNB is connected without a switch in between.
The value 2 selects the second position on the DiSEqC switch etc.
The value is stored in the already existing but previously unused field dvb_diseqc_type
in table capturecard; there is no schema upgrade needed for this feature.