MythExternRecorder: Add optional [VARIABLES] in config file. Allow TUNE cmd in channels.conf
If the config file contains a [VARIABLES] section, any %KEY% variables found in any of the commands will be replace with the given value.

In the channels configuration file, if a
TUNE='cmd'
and/or
ONSTART='cmd'
is provided for a channel, it will be used instead of what is in the main config file.