Segmentation fault in mythfrontend guide when there are no channels
Avoid accessing the first channel if the number of channels is 0 when showing the guide grid.
The code to compute the m_channelCount is now simplified and this solves the problem.
However, it remains a mystery what the purpose of the original code was.
Thanks to ulmus-scott for providing a first version of this fix in PR#648.