Fix mythfrontend crash when playing DVB radio
Check if there is a buffer pointer in the letterbox detection.
It is not clear if the letterbox detection should be active when
playing radio and if letterbox detection it is active it should
not be called with an invalid frame but anyway checking for null
pointers is always OK.

Thanks to Steve Evans / TheGorbag.

Refs #689