cppcheck: Fix "dereference of an invalid iterator" warnings.
These loops currently iterate from "end()-1" to "begin()-1", which
attempts to reference the nonexistent iterator that comes before
"begin()". This code is a crude attempt to iterate lists in the
reverse order (although reverse iterators may not have existed at the
time the code was written).  Converting these loops to use reverse
iterators eliminates the errors.

The validity of this solution was checked by temporarily using both
solutions, and printing out item address to confirm that both loops
walked the list in the same order.