Send REC_PENDING event before REC_STARTED event
Take the prerollseconds into account when determining the moment
to send the REC_PENDING event. This was never done and as a result
the REC_PENDING event was never sent before the recording started
when there were more than 120 prerollseconds configured.
Also, due to updating the recstartts from the scheduled start time
to the actual start time, the unique key identifying the recording
did change. This then caused an additional REC_PENDING event to be
sent immediately after the REC_START event. This has been fixed by
checking that there is still time left before the recording will start.

Thanks to PhilB for reporting the issue and to paulh for the first analysis.