HDHomeRun DVB-T/T2 tuning command
For HDHomeRun DVB-T/T2 tuners, use the bandwidth and the modulation system in the tuning command.

Obtain signal bandwidth, 6, 7 or 8MHz, and the DVB-T/T2 modulation system, 'DVB-T' or 'DVB-T2', from
table dtv_multiplex in the database instead of from the tuner type. This allows a multiplex to be tuned
with the correct modulation string.
The code for this was already present but not correct and this caused always the "auto" mode to be used.

An example of an explicit tuning command for a DVB-T2 signal with 8MHz bandwidth on a frequency
of 570MHz is "t8dvbt2:570000000".

To get the list of supported modulation modes use the following command
./hdhomerun_config FFFFFFFF get /sys/features
where FFFFFFFF is the 8-digit hex identfier of the HDHomeRun.

The tuning command for HDHomeRun DVB-C has already been changed along the same lines in commit b6b3947 on Oct 29, 2019.
For DVB-C this has reduced the time needed to obtain a lock and it has increased the reliability.
An explanation for this is that with the "auto" mode, a HDHomeRun tries all possible combinations of bandwidth
and modulation modes that are supported for both DVB-T/T2 and DVB-C. This can take a long time.
The expectation is that the DVB-T/T2 tuning will also be quicker and more reliable but this has not yet been proven.