Fix pointer to integer (to pointer) casting warnings in audioconvert.
Don't cast a pointer to an integer, add a number to it, and then cast
it back to a pointer.  This transformation is not transparent to the
compiler, and it can lose potential optimizations chances.  Instead
cast it to a pointer to a uint8_t (width of one byte) and then do the
addition directly to the pointer.