mythlink including deleted recordings
Fix for mythlink.pl so that it does not create links anymore to deleted
items (or more specifically, items in the Deleted recording group),
which is likely not desired behavior.

Patch supplied by Karl Neuman at Nov 25, 2013, 6:49 AM as described in
https://lists.archive.carbon60.com/mythtv/users/558623

Thanks also to Robin Gilks for reporting the issue and
for Stephen Worthington and John Pilkington for pointing to the
solution and confirming the correctness.