cmake: Compile dvbdescriptors.cpp on win32.
By default, the iconv headers make the second argument to iconv() a
"const char **". Unfortunately the compiler will not implicitly cast
the "char **" used by this file into that type.  Add the define so
that the type on windows is "char **" to match our usage.  This define
has no effect on other systems.