mythtv-setup hanging at Video Sources
On Ubuntu 22.10 mythtv-setup hangs in Video Sources when creating a new Video source.
The problem has been traced to running tv_find_grabbers by MythSystemLegacy.
The fix consists of adding the kMSDontDisableDrawing flag to the invocation of MythSystemLegacy.
It is not clear why this problem only appears on Ubuntu 22.10 and not on other distro's.

Refs #690