HDHomeRun DVB-T/T2 tuning command
For HDHomeRun DVB-T/T2 tuners, use the bandwidth but not the modulation system in the tuning command.
For example, use "autot8:570000000" for a 570MHz signal.
Previously, instead of "auto" the modulation system was specified.
This led to issues when scanning for channels because then the modulation system is not yet known; it happens
on some systems that Using "dvbt" as modulation system can give a tuning error on a dvb-t2 multiplex.
This partially reverts commit fff7d84c0ea2e65cc7f95a711656a5dd889b104b where both
bandwidth and modulation system were added to the tuning command.