Square pixels for all HD/FHD/UHD displays
For the GUI of MythTV, use a fixed pixel aspect ratio of 1.0 for all displays
that have a vertical resolution of at least 720 pixels.
Previously, the pixel aspect ratio was always computed from both the resolution and
the dimensions of the display as reported by the EDID when both are available.
This solves GUI issues when the display dimensions as reported by the EDID is not correct.
Note that for video playback the setting "Screen aspect ratio" is used.