cmake: Fix warning about  narrowing in an initializer list.
The compiler is complaining that it can't narrow a value from type
"long long" to "long" in an initializer list.  Given that this code is
using milliseconds, then by definition the return value of the count()
function contains at least 45 bits, and is probably always a int64_t.
The timestamp tv_nsec field (a long) can be either 32 or 64 bits
depending on the definition of "long" on the platform.  Add an
intermediate variable so that the compiler can perform the narrowing
outside of an initializer, and thus doesn't complain.