Fix detection of optical drives
Checking for mount points is not reliable to determine if a block device
is an optical drive. This commit uses the 'MediaCompatibility' property
of the 'org.freedesktop.UDisks2.Drive' DBUS interface to check if the
drive supports optical media.
Also changes the function name in the log messages to match the function
they are called from.

(cherry picked from commit 998669de6114706d0459214ecc030aed7489afab)