Remove std::move from Service API macro
std::move is dangerous to use without careful analysis. It causes the
original value to be cleared. This resulted in a problem with
Myth/GetBackendInfo and may cause unexpected problems with other API
methods. Removing it from the macro removes it from all API value
setting and causes an assignment to be done instead.

This commit also reverts commit c92362e , which was a workaround fix
for Myth/GetBackendInfo.

Fixes #739