Remove the qjsonwrapper (qt4->qt5) bridge code.
The only object that used this code for serialization/deserialization
was the LoggingItem object.  This was needed when the logging server
was built as a separate executable and the LoggingItem needed to be
passed from one program or machine to another.

The client/server part of the logging server code was removed in 2018
(commit 773ede8298) so all the manipulation is done inside a single
program, but the serialization/deserialization code was left in place.
This commit doesn't remove that code, it just changes the
serialization/deserialization functions. These functions are not
compatible with the previous implementation, but they do not need to
be as the JSON produced is internal to a single program.

The serialization/deserialization is still unnecessary and should be
remove in a future commit.