Remove support for logging to the database.
This commit removes the support that allows logging messages to be
written to the database in addition to being logged to the console or
sent to logger/syslog.  Several classes are removed from the logging
server code.  It also drops the "logging" table from the database, and
removes the background task that periodically cleaned up the that
table.