Fixes to firewire class
In order to add firewire support to service API
Capture/GetCaptureDeviceList, make the FirewireDevice class public and
fix the header includes to use the full path so the header can be
included from a different directory.

Fix the test data: the member variables now have m_ prefix but the test
data did not.