HDHomeRun full MPTS recording
Change the way a full MPTS recording is implemented.
Pid 0x2000 is used to request a full transport stream.
This is how it is done in /dev/dvb drivers and the HDHomeRun
does accept that but it does then give an error message.
The range of pids is now changed to 0x0000-0x1FFF when 0x2000
is in the list of pids.