HDHomeRun Channel Import
Add the option "HDHomeRun Channel Import" to get the channel list
directly from the HDHomeRun. The HDHomeRun does a channel scan
all by itself and the list of channels can now be read by mythtv-setup.
Also the http url for each channel is read and stored in the database
although this is not yet used in MythTV.
There is enough information read from the HDHomeRun to fill database tables
channel and dtv_multiplex so that recording can be done with libhdhomerun.
This feature has only been tested in The Netherlands and it is expected
that it works with all European DVB-T/T2 and DVB-C signals.
Support for reading USA/ATSC channel lists is to be added.
The actual code is shamelessly copied from the channel import for the VBox
and adapted where needed.