Show "Input name" field only for analog inputs
In mythtv-setup, Input Connections configuration, show the
"Input name" selection only for cards with analog inputs.
This field is used to select for example between Composite,
S-Video or Component Video when a card has different inputs.
Note that for digital cards the same database field is used
to store the Delivery System.