External subtitles now use the same code path as embedded subtitles.
Read AVSubtitles from the external file and enqueue them directly onto
the AVSubtitle queue.  This is the same queue used for AVSubtitles
embedded directly into in a video file.

External subtitles will no longer appear as white letters on a black
background, but will be styled the same way as embedded subtitles.
The default for this appears to be white letters with a slight drop
shadow overlaid on the picture (without the black background box).

This eliminates a bunch of code that was used to read the text out of
the AVsubtitle, queue it onto a separate text only queue, and prints
it onto the screen.