Backend web-server only mode without database for setup.
The backend can start without a database, so that the
database can be set up using the web app.

If any of these conditions exist at backend startup,
the backend will start in a web-server only mode:

- The config directory (normally .mythtv) does not exist.
- Time zone tables are not installed.
- The database cannot be connected after 10 attempts.
- Host address backend setup is not done.

The backend will no longer prompt for entry of database
credentials through the terminal. These must be done
through the web app or manually with config.xml.