tidy: Fix two "undefined binary operator result" warnings.
Tidy is warning that "the left operand of '<' is a garbage value" in
both instances of the visualization code.  Convert this C vector into
a C++ vector and initialize it as part of setting it to the right size.

(cherry picked from commit 84b3222aefdd0adc0dad5b8aa28f2b875ea6e8fe)