IPTV URL in Channel Editor
In the mythtv-setup Channel Editor add inspection and modification
of field url in table iptv_channel. This field contains the URL for
streaming as used by the IPTV capture card.
The URL data can be obtained with a "M3U Import" from an M3U file defined in
the IPTV capture card or with a "HDHomeRun Channel Import" where an XML file
with channel data from the HDHomeRun is loaded.