Fix parsing VBI teletext PES packets
Fix parsing VBI PES packets with teletext menu pages and with teletext
subtitling. These PES packets contains of a number of teletext packets
as described in ETSI EN 301 775 V1.2.1 (2003-05).
The problem with the previous implementation was that it dit not correctly
parse stuffing teletext packets that appear if needed at the end, to fill up
the PES packet. This then caused an endless stream of error messages
when VBI logging was enabled.