Terminate Live TV channel input with SELECT (v2)
In Live TV another channel can be selected by typing in the channel number.
The selection takes effect five seconds after the first character has been entered.
The channel number input can now be terminated by a SELECT action (the Enter key).

This is the second version; the SELECT action is only done when there is
a channel number entered, otherwise the SELECT action is passed on.