Add parentheses to clarify validity test in libdvdnav.
The compiler is issuing this set of warning:

    vmget.c:196:8: warning: logical not is only applied to the left
    hand side of this bitwise operator [-Wlogical-not-parentheses]

    note: add parentheses after the '!' to evaluate the bitwise
    operator first

    note: add parentheses around left hand side expression to silence
    this warning

This code looks like it is trying to test whether a single bit in an
integer is set, so wrap the entire expression in parentheses.

Note: The original libdvdnav sources were checked for their solution
to this error, but this function has never existed there.  Its a
MythTV addition.