Fix error in power search / custom record SQL validation
When creating a Custom record rule, the SQL validation failed to detect
an error with title = 'Some Value'. The test query only contained
two tables but the actual query used has three tables.

Also changed to return the error text when the service API V2 is used to
create the Power Search rule.