subtitles: Stop requesting external subs after FFmpeg returns empty.
After the last external subtitle has been read, FFmpeg will set the
packet buffer to nullptr and return an error code.  Set a flag when
FFmpeg returns an error and stop calling every "pulse" to see if a new
subtitle is available.  If the user seeks, assume that it could be a
backwards seek and start checking for subtitles again.  Either a
subtitle will be returned on the next call to FFmpeg, or the code will
go back into "no more subtitles" mode.