Fix heap-use-after-free in TVRec with Live TV
When playing Live TV continuously on the same channel, there is
a moment when according to the guide one program ends and the next
program begins. At this moment the Live TV recording is stopped and
a new recording is started. There is small time period in between
in which there is no valid program info and an update of the program
info from another thread in that period results then in access
of already deleted program info instance.
Fixed by using the m_stateChangeLock to lock the complete sequence.

Refs #768
Refs #769

Signed-off-by: Klaas de Waal <klaas@kldo.nl>