mythfrontend: Prevent inactive override rules
When a user creates an override for an episode, there
is an option to set it inactive. If the user sets
it inactive, the override is ignored by the scheduler.

If there is an inactive override, the user interface
shows an option to create an override. If they try
to create another override, saving the override fails
dues to a database duplicate error, and the error is
not reflected back to the user. The user assumes the
override was successful but it fact it was lost.

On the other hand, when there is an active override,
the user interface shows options to modify or delete
the override, and there is no danger of creating a
duplicate.

I solve this by disabling the ability to set overrides
as inactive. Aside from creating this potential error,
there seems no legitimate use for an inactive override.

Note: you can still make an inactive override by applying
an inactive template to an override. This fix does not
address that possibility.