Fix potential null pointer dereference in channelscan_sm.
Recent changes to Qt6 or the compiler have produced new null pointer
dereference warnings.  Apparently every call to a begin()/find()
function has a code path that can return the end() iterator (which
makes sense), and dereferencing the end() iterator in Qt6 is a null
pointer dereference.  This means that any call to a function returning
an iterator must be followed by a validity test before dereferencing
that iterator.

Rewrite the initial value checks so the code has access to the
iterators.  Use the already validated iterator for the rest of the
function instead of calling find() a second time.