Run UpdateChannelGroups during startup of master backend
web-app setup can add, modify or delete channels. The changes need to be
reflected in the automatic channel groups. To ensure that this happens,
UpdateChannelGroups is run during every start of a master backend. It is
run after the scheduler is started. This ensures that running it does not
cause a delay in starting recordings.