MythMusic: "Play Now" any tracks from the music browsers
ref: https://forum.mythtv.org/viewtopic.php?p=25164#p25164

Spontaneous listeners want to just play an album or song now and care
less about any playlist.  This adds a simple "Play Now" to the menus
which is identical to "Add Tracks" except that it immediately starts
playing the additions.

This required changing the working list to allow duplicates as you
need the whole album now even if you played just part of it yesterday.
Of course removing a track removes all occurances of it.