Skip SeekReset before scanning streams again
In StreamChangeCheck, skip SeekReset before ScanStreams is called.
The SeekReset discards processed video frames and input buffers and this causes a part
of the video not being played plus a video freeze of about half a second.
It looks like this code is only called when there is a change in the streams after a new PMT
has been received and this happens only incidentally with streams from Finnish broadacaster YLE.
Skipping the SeekReset gives smooth playback on the YLE streams when there is a stream change.
A call to StreamChangeCheck can also be initiated from FFmpeg from libavformat/mpeg.c but I have not
been able to find a recording or a video other than the YLE recordings that does trigger this.
There is now a log message if the skip does happens just in case this might cause a regression.
For more details see commit 4c49293a5b8a531603a348fccbf90816c9d939cb.