MythMusic: make it easier to find the music library and get started
"Playlist Editor" is renamed to "Browse Music Library" on the
menu. Reasoning:

* the playlist itself can already remove and move tracks, part of editing
* the playlist editor is for only adding (or replacing)
* but it is more about browsing the music collection in useful ways
* new users might not realize their library is in an "editor"
* after browsing, the menu makes it clear how to add tracks from there:

The empty list replaces "You haven't selected any tracks to play" with
"Press M (Menu) to add and play tracks" giving a clue how to start.

Internally the code still calls this view the playlist editor - only
the menu item is changed.  We can update the manual in a few places
and note that these are synonyms or something like "Browse Music
Library" was the "Playlist Editor" prior to MythTV 34.