MythMusic: "Play Now" any tracks from the music browsers (#774)
* MythMusic: "Play Now" any tracks from the music browsers

ref: https://forum.mythtv.org/viewtopic.php?p=25164#p25164

Spontaneous listeners want to just play an album or song now and care
less about any playlist.  This adds a simple "Play Now" to the menus
which is identical to "Add Tracks" except that it immediately starts
playing the additions.

This required changing the working list to allow duplicates as you
need the whole album now even if you played just part of it yesterday.
Of course removing a track removes all occurances of it.

* MythMusic: enter browser at Albums rather than All Tracks

* MythMusic Play Now: checked tracks toggle as expected

* MythMusic: Play Now after Add Tracks in editor menu

* MythMusic: Cancel shuffles and repeats to Play Now

* MythMusic s/Play Now.../Play Now/g

* MythMusic: add Play Now to Playlist menus

* MythMusic: move re-added tracks to the end of the list

So that adding an album today doesn't miss tracks added yesterday.

* MythMusic: enter the music browser where we left last time

* MythMusic: Merge Play Now into Playlist Editor to be less confusing

as the 2 features are nearly identical.  A new menu option toggles
"Prefer Play Now" or "Prefer Add Tracks" which orders them on the menu
and defines the action for selecting a single track.

* MythMusic: support play/add preference also in music search

* MythMusic: fix bug that switching from Gallery to Tree broke repaints

upon escape back to the playlist.  I don't know what these disconnects
were for, but removing them restores updates to the parent window.

* MythMusic: make it easier to find the music library and get started

"Playlist Editor" is renamed to "Browse Music Library" on the
menu. Reasoning:

* the playlist itself can already remove and move tracks, part of editing
* the playlist editor is for only adding (or replacing)
* but it is more about browsing the music collection in useful ways
* new users might not realize their library is in an "editor"
* after browsing, the menu makes it clear how to add tracks from there:

The empty list replaces "You haven't selected any tracks to play" with
"Press M (Menu) to add and play tracks" giving a clue how to start.

Internally the code still calls this view the playlist editor - only
the menu item is changed.  We can update the manual in a few places
and note that these are synonyms or something like "Browse Music
Library" was the "Playlist Editor" prior to MythTV 34.