tidy: Disable a couple of erroneous warning messages.
1) The MythDRMDevice::GetModes can't possibly be converted to static
as it references member variables.

2) The PlaybackProfileConfig::swap function makes no attempt to
prevent throwing an exception.

3) The TruncateThread object in MainServer::HandleDeleteFile should be
automatically deleted by QRunnable once it finishes executing.  There
is no memory leak.

4) The CustomPriority::loadData function doesn't lose the object
pointed to by the item variable 'item'.  When MythUIButtonListItem
creates an item it is added to the list passed as the first argument,
so can still be referenced/deleted from there.  There is no memory
leak.