mythbackend_main_helpers: prevent repeated systemd restarts
When the backend is running in the 'Web APP only mode', and is started by
systemd with Type=notify, it was repeatedly getting restarted because the
READY=1 sd_notify hadn't been sent.

Additionally, update comments: port now 6544, not the 6744, use the name
Web App, replace references to mythtv-setup with Web App.

Once running in the 'Web App only mode', systemctl status mythtv-backend.service
will report the Status like this:

    Status: "Started in Web App only mode"