Save total frame count for video files
When exiting the media player, check if the current
video file alreadey has a total frame count saved
within the filemarkup table. If it is not yet present,
then save the total frame count from the player.
This way the watched progress does not need extra
tooling to provide the total frame count of a video file.