MythMusic: Rename removeDuplicateTracks to removeItemsFromList (#800)
* MythMusic: Rename removeDuplicateTracks to removeItemsFromList

to more accurately describe what this function does.  The end result
is no duplicates but that is due to how this function is called.
Tracks are added by using this function to first remove the new adds
from the prior list, then concatenating the modified result and new
list.  The result is that any previously selected tracks are moved to
the end of the list in order as needed by Play Now and Add Tracks.

* MythMusic: remove accidental tab from playlist code