OSX: Fix FFMPEG Xcode15 linker issue
With the introduction of Xcode 15, Apple has moved to a new linker and
introduced the "-ld_classic" linker flag as a work around. Without this
flag, FFMPEG fails to compile on systems with Xcode 15 (i.e. Sonoma).

This workaround may need to be removed at the next FFMPEG sync or when
Apple deprecates the flag.