v2 api: GetVideoList: cater for multiple sort fields
The service now can take any number of sort fields separated by commas.
Also each sort field can optionally have a space and ASC or DESC to
individually control ascending or descending for that sort field. The
global Descending parameter is the default for fields that do not
individually specify ascending or descending.