web app: Fix sort order of recordings and videos, fix transcode menu
If recordings are sorted by clicking on a column,  make sure that there
is a sensible secondary sort key. For example if sorting by title,
episodes must be sorted in airdate order. If sorting by rec group, shows
must be in title order.

Change menu item for Transcode to "Transcode with Default profile".