EXIF tag 0x9286 UserComment can contain garbage
In the photo gallery view and in the thumb view there can be
comment presented when present in the EXIF data.
On pictures made with a Samsung phone and with a Canon camera
the UserComment, as read with the EXIF library, can contain
non-printable characters with numeric value -1.
All non-printable characters are now removed from the UserComment
before the comment is presented.