Use the metadata grabber cache when trying to find a grabber.
The s_grabberAge variable starts in an invalid state.  The old code
didn't call the GetList() because the s_grabberAge variable was
invalid, and therefore the variable was never set to a valid value.
Remove this extra (broken) cache validity test and let the GetList
function handle whether/when it updates the cache.